Updated version of the subgraph plus redireciton for adding lore
