+++ pages/lore/[wizardId]/[page].tsx
@@ -1,5 +1,5 @@
 import Layout from "../../../components/Layout";
-import { GetServerSidePropsContext, GetStaticPropsContext } from "next";
+import { GetStaticPropsContext } from "next";
 import Book from "../../../components/Lore/Book";
 import client from "../../../lib/graphql";
 import { gql } from "@apollo/client";
@@ -12,6 +12,11 @@ import dynamic from "next/dynamic";
 import { IPFS_SERVER } from "../../../constants";
 import productionWizardData from "../../../data/nfts-prod.json";
 import { flatMap } from "lodash";
+import {
+  getCurrentWizardData,
+  getPreAndNextPageRoutes,
+} from "../../../components/Lore/loreSubgraphUtils";
+
 const wizData = productionWizardData as { [wizardId: string]: any };
 
 const WizardMapLeaflet = dynamic(
@@ -36,146 +41,6 @@ const LorePage = ({
   );
 };
 
-const COMMON_LORE_FIELDS = `
-  id
-  index
-  creator
-  tokenContract
-  loreMetadataURI
-  parentLoreId
-  tokenId
-  struck
-  nsfw
-  createdAtBlock
-  createdAtTimestamp
-`;
-
-async function getPreviousAndNextPagedata(
-  wizardNum: number,
-  leftPageData: any,
-  rightPageData: any
-) {
-  let queryString = "";
-
-  // Add previous page if we in middle of book
-  if (leftPageData) {
-    queryString = `${queryString}
-      prevRightPage: lores(first: 1, where: {id_lt: "${leftPageData.id}", struck: false, nsfw: false}, orderBy: id, orderDirection:desc) {
-        id
-      }
-    `;
-  } else {
-    queryString = `${queryString}
-      prevRightPage: lores(first: 1, where: {id_lt: "wizards-${wizardNum
-        .toString()
-        .padStart(
-          5,
-          "0"
-        )}-0000", struck: false, nsfw: false}, orderBy: id, orderDirection:desc) {
-        id
-      }
-    `;
-  }
-
-  if (rightPageData) {
-    queryString = `${queryString}
-      nextLeftPage: lores(first: 1, where: {id_gt: "${rightPageData.id}", struck: false, nsfw: false}, orderBy: id, orderDirection:asc) {
-        id
-      }
-    `;
-  } else if (leftPageData) {
-    queryString = `${queryString}
-      nextLeftPage: lores(first: 1, where: {id_gt: "${leftPageData.id}", struck: false, nsfw: false}, orderBy: id, orderDirection:asc) {
-        id
-      }
-    `;
-  } else {
-    queryString = `${queryString}
-      nextLeftPage: lores(first: 1, where: {id_gt: "wizards-${wizardNum
-        .toString()
-        .padStart(
-          5,
-          "0"
-        )}-0000", struck: false, nsfw: false}, orderBy: id, orderDirection:asc) {
-        id
-      }
-    `;
-  }
-
-  const { data: prevAndNextPageData } = await client.query({
-    query: gql`
-        query WizardLore {
-            ${queryString}
-        }
-    `,
-  });
-
-  const nextLeftPageData = prevAndNextPageData?.nextLeftPage?.[0] ?? null;
-  const prevRightPageData = prevAndNextPageData?.prevRightPage?.[0] ?? null;
-
-  let previousWizardPageCount = 0;
-
-  if (prevRightPageData) {
-    // get previous wiz page count we can figure out the page number
-    const prevRightPageWizardNum = parseInt(
-      prevRightPageData.id.match(/^(\d+)-(\d+)$/) ?? "0"
-    );
-
-    if (prevRightPageWizardNum !== wizardNum) {
-      const { data: prevWizardPageCount } = await client.query({
-        query: gql`
-            query WizardLore {
-                lores( where: {wizard: "${prevRightPageWizardNum}", struck: false, nsfw: false}) {
-                    id
-                }
-            }
-        `,
-      });
-
-      previousWizardPageCount = (prevWizardPageCount?.lores ?? []).length;
-    }
-  }
-  return {
-    nextLeftPageGraphData: nextLeftPageData,
-    prevRightPageGraphData: prevRightPageData,
-    previousWizardPageCount,
-  };
-}
-
-async function getCurrentWizardData(
-  rightPageNum: number,
-  wizardNum: number,
-  leftPageNum: number
-) {
-  let currentWizAndPagequeryString = `
-      rightPage: lores(skip: ${rightPageNum}, first: 1, orderBy: id, orderDirection: asc, where: {wizard: "${wizardNum}", struck: false, nsfw: false}) {
-        ${COMMON_LORE_FIELDS}
-      }
-  `;
-
-  // Add left page to query unless we on very first page (instead wiz is shown)
-  if (leftPageNum >= 0) {
-    currentWizAndPagequeryString = `${currentWizAndPagequeryString}
-         leftPage: lores(skip: ${leftPageNum}, first: 1, orderBy: id, orderDirection: asc, where: {wizard: "${wizardNum}", struck: false, nsfw: false}) {
-            ${COMMON_LORE_FIELDS}
-          }
-     `;
-  }
-
-  const { data: currentWizLoreData } = await client.query({
-    query: gql`
-        query WizardLore {
-            ${currentWizAndPagequeryString}
-        }
-    `,
-  });
-
-  return {
-    leftPageGraphData: currentWizLoreData?.leftPage?.[0] ?? null,
-    rightPageGraphData: currentWizLoreData?.rightPage?.[0] ?? null,
-  };
-}
-
 async function fetchLoreMetadata(loreMetadataURI: string | null): Promise<any> {
   if (!loreMetadataURI) return null;
 
@@ -188,6 +53,7 @@ async function fetchLoreMetadata(loreMetadataURI: string | null): Promise<any> {
     return null;
   }
 
+  //TODO: retries?
   return await fetch(ipfsURL).then((res) => {
     if (res.ok) {
       return res.json();
@@ -248,16 +114,6 @@ export async function getStaticProps(context: GetStaticPropsContext) {
     };
   }
 
-  const {
-    nextLeftPageGraphData,
-    prevRightPageGraphData,
-    previousWizardPageCount,
-  } = await getPreviousAndNextPagedata(
-    wizardNum,
-    leftPageGraphData,
-    rightPageGraphData
-  );
-
   let leftPage: IndividualLorePageData;
 
   if (leftPageGraphData) {
@@ -287,34 +143,12 @@ export async function getStaticProps(context: GetStaticPropsContext) {
     };
   }
   rightPage.pageNumber = rightPageNum;
-
-  // Figure out previous route
-  let previousPageRoute = null;
-  if (prevRightPageGraphData) {
-    // Previous page could be this wizards or previous wizard's last
-    const previousIdMatcher = prevRightPageGraphData.id.match(
-      /^wizards-(\d+)-(\d+)$/
-    );
-    const previousPageWizNum = parseInt(previousIdMatcher?.[1] ?? "0");
-    const previousPagePageNum =
-      previousPageWizNum === wizardNum
-        ? leftPageNum - 1
-        : previousWizardPageCount > 0
-        ? previousWizardPageCount - 1
-        : 0;
-    previousPageRoute = `/lore/${previousPageWizNum}/${previousPagePageNum}`;
-  }
-
-  // Figure out next route
-  let nextPageRoute = null;
-  if (nextLeftPageGraphData) {
-    const nextIdMatcher = nextLeftPageGraphData.id.match(
-      /^wizards-(\d+)-(\d+)$/
-    );
-    const nextPageWizNum = parseInt(nextIdMatcher?.[1] ?? "0");
-    const nextPagePageNum = nextPageWizNum === wizardNum ? pageNum + 2 : 0;
-    nextPageRoute = `/lore/${nextPageWizNum}/${nextPagePageNum}`;
-  }
+  let { previousPageRoute, nextPageRoute } = await getPreAndNextPageRoutes(
+    wizardNum,
+    pageNum,
+    leftPageGraphData,
+    rightPageGraphData
+  );
   console.log({
     leftPage,
     rightPage,
@@ -339,8 +173,9 @@ export async function getStaticPaths() {
   const { data } = await client.query({
     query: gql`
       query WizardLore {
-        wizards(orderBy: id) {
-          id
+        loreTokens(orderBy: tokenId) {
+          tokenContract
+          tokenId
           lore(
             where: { struck: false, nsfw: false }
             orderBy: id
@@ -353,19 +188,20 @@ export async function getStaticPaths() {
     `,
   });
 
+  console.log(data);
+
   //Note: its so annoying NextJs doesn't let you pass extra data to getStaticProps so now we fetch inside there too sigh... https://github.com/vercel/next.js/discussions/11272
-  const paths = flatMap(data.wizards, (wizardData: any) => {
-    return (wizardData?.lore ?? []).map((loreData: any, index: number) => {
+  const paths = flatMap(data.loreTokens, (loreTokenData: any) => {
+    return (loreTokenData?.lore ?? []).map((loreData: any, index: number) => {
       return {
         params: {
-          wizardId: wizardData.id,
+          wizardId: loreTokenData.tokenId,
           page: (index % 2 === 0 ? index : index + 1).toString(),
         },
       };
     });
   });
-  // paths.filter((path) => parseInt(path.page) % 2 === 0); //TODO: this omits final lore page if its only left
-  // console.log(paths);
+
   return {
     paths: paths,
     fallback: "blocking",
